HIVE-5799 : session/operation timeout for hiveserver2 (Navis, reviewed by Brock Noland, Lars Francke and Lefty Leverenz)
git-svn-id: https://svn.apache.org/repos/asf/hive/trunk@1621912 13f79535-47bb-0310-9956-ffa450edef68

+++ ql/src/java/org/apache/hadoop/hive/ql/exec/AutoProgressor.java
@@ -44,7 +44,7 @@ public class AutoProgressor {
   // Name of the class to report for
   String logClassName = null;
   int notificationInterval;
-  int timeout;
+  long timeout;
   Reporter reporter;
 
   class ReporterTask extends TimerTask {
@@ -116,7 +116,7 @@ public class AutoProgressor {
    * @param timeout - when the autoprogressor should stop reporting (in ms)
    */
   AutoProgressor(String logClassName, Reporter reporter,
-      int notificationInterval, int timeout) {
+      int notificationInterval, long timeout) {
     this.logClassName = logClassName;
     this.reporter = reporter;
     this.notificationInterval = notificationInterval;
